triton_client/requirements.txt
attrdict==2.0.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
cycler==0.10.0
distro==1.5.0
docker==5.0.0
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.1.0
grpcio==1.37.1
grpcio-tools==1.37.1
httplib2==0.19.1
idna==2.8
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.4.1
numba==0.53.1
numpy==1.20.2
pdfkit==0.6.1
Pillow==8.2.0
prometheus-client==0.10.1
protobuf==3.16.0
psutil==5.8.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-rapidjson==1.0
PyYAML==5.4.1
requests==2.25.1
requests-unixsocket==0.2.0
six==1.14.0
tritonclient
urllib3==1.25.8
websocket-client==0.59.0
zope.event==4.5.0
zope.interface==5.4.0
